Fixes Android bug when called before application.android.context is available.
When before application.android.context is available, such on startup of app, this module will fail with "Error in appversion.getVersionName: TypeError: Cannot read property 'getPackageManager' of undefined". This changes so that the *current* application.android.context is used when called, instead of the application.android.context it got when loading the app.